Moved getTLE and isSunlit to util
